Made strings solely a pointer on the stack
Strings now behave like bona fide objects as far as assignment
goes; a copyStr function will be added to deep-copy strings for the
purpose of using them like primitives. This is to avoid the issue
of when to free string literals (which need to be malloced so that
they can be returned from functions properly).